+++ CHANGES.md
@@ -1,6 +1,8 @@
 master (in development)
 =======================
 
+- Add `flycheck-pos-tip-timeout` [GH-9]
+
 0.1 (Nov 16, 2015)
 ==================